// public/app/types/devicemanagement/group/groupuser.ts
export const groupUsersPageLimit = 50;

export interface GroupUser {
  id: number;
  login: string;
  email: string;
  phone: string;
  name: string;
  role: string;
}

export interface GroupUsersState {
  groupUsers: GroupUser[];
  searchQuery: string;
  searchPage: number;
  groupUsersCount: number;
  hasFetched: boolean;
}